requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=6.0.0

# API connectivity
requests>=2.26.0
websocket-client>=1.2.1

# Command line interface
click>=8.0.0

# Data visualization (optional, for backtesting reports)
matplotlib>=3.4.0
seaborn>=0.11.2

# Notification services
discord.py>=2.0.0

# Testing
pytest>=6.2.5

# Utility
python-dotenv>=0.19.0